src/backend/requirements.txt
aiohttp
aiofiles
azure-ai-documentintelligence == 1.0.0
azure-ai-inference == 1.0.0b6
azure-identity
tqdm 
python-dotenv
requests
openai
instructor
pydantic
pillow
rich
azure-storage-blob == 12.24.0
--extra-index-url https://pkgs.dev.azure.com/azure-sdk/public/_packaging/azure-sdk-for-python/pypi/simple/
azure-search-documents==11.6.0a20250505003
pyjwt